requirements.txt
fastapi
uvicorn
python-docx
cohere
pymupdf
pandas
openpyxl